add verifyPrecedingChannel and move stuff to functions.js
